Add option for creating code-signing certificates.
This commit adds support for signing certificates with the codeSigning
extended key usage extension. Tried to make this as non-invasive as
possible by adding a new function to create host certificates api with
an optional array of []x509.ExtKeyUsage. This should allow adding
support for other extended key usages in the future. For now, just added
a simple boolean flag to the sign command with turns this on.