Expand on permissions error message.
If permissions are not as expected by certstrap it will give a very
terse "permission denied" message that gives no indication of what the
error is or how to address it.  This small PR expands the error message
to explain with which file the permissions problem lies, and why the
problem arose.  This allows users to be able to address the issue.